Migrated SVGElement to ES6 pattern.
